Pindora_shield/requirements.txt
# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.34.0
pydantic==2.10.3

# Machine Learning & Deep Learning
torch==2.5.1
pytorch-lightning==2.4.0
catboost==1.2.7
scikit-learn==1.6.0
optuna==4.1.0

# Chemistry & Molecular Processing
rdkit==2024.3.6

# Data Processing
numpy==2.0.2
pandas==2.2.3
joblib==1.4.2

# API & HTTP
requests==2.32.3
python-dotenv==1.0.1

# Groq (replaces Azure OpenAI)
groq>=0.9.0

# Additional Utilities
tqdm==4.67.1